Refactor works
