+++ src/common/enclave_interface_types.h
@@ -92,10 +92,3 @@ constexpr char const* create_node_result_to_str(CreateNodeStatus result)
     }
   }
 }
-
-enum StartType
-{
-  Start = 1,
-  Join = 2,
-  Recover = 3,
-};
