+++ src/ds/cli_helper.h
@@ -133,35 +133,11 @@ namespace cli
     return option;
   }
 
-  struct SizeString
+  inline static size_t convert_size_string(std::string input)
   {
-    size_t value;
-
-    SizeString() = default;
-    SizeString(size_t val) : value(val) {}
-
-    bool operator==(const SizeString&) const = default;
-
-    inline operator size_t() const
-    {
-      return value;
-    }
-  };
-
-  inline void to_json(nlohmann::json& j, const SizeString& str)
-  {
-    // CLI11 does not provide convenient way to convert back to size string
-    throw std::logic_error(
-      "to_json() method for SizeString type is not implemented");
+    size_t ret = 0;
+    CLI::AsSizeValue(false)(input); // Parse both all values as multiple of 1024
+    assert(CLI::detail::integral_conversion(input, ret));
+    return ret;
   }
-
-  // TODO: This should only be pulled by host
-  inline void from_json(const nlohmann::json& j, SizeString& str)
-  {
-    auto val = j.get<std::string>();
-    CLI::AsSizeValue(false)(val); // Parse both all values as multiple of 1024
-    assert(CLI::detail::integral_conversion(val, str.value));
-    LOG_FAIL_FMT("from_json size str: {}", str.value);
-  }
-
 }
