+++ src/enclave/enclave.h
@@ -24,6 +24,12 @@
 
 namespace enclave
 {
+  // TODO: Move elsewhere?
+  inline void from_json(const nlohmann::json& j, SizeString& str)
+  {
+    str = j.get<size_t>();
+  }
+
   class Enclave
   {
   private:
